requirements/common.txt
asgi-redis==0.14.1
celery==4.0.0
Django==1.10.1
django-countries==4.0
django-crispy-forms==1.6.1
django-autoslug==1.9.3
django-model-utils==2.5.2
psycopg2==2.6.2
setuptools==28.8.0
django-oauth-toolkit
djangorestframework
markdown
django-filter
django-cors-middleware
django_oauth2_provider
redis==2.10.5
requests==2.4.3
pyyaml
